% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean Column Names
A wrapper around janitor::clean_names() that converts messy column
names into snake_case.}
\usage{
clean_names(df)
}
\arguments{
\item{df}{A data.frame or tibble.}
}
\value{
A data.frame/tibble with cleaned column names and an extra
  S3 class \code{tidycleanr_cleaned}.
}
\description{
Clean Column Names
A wrapper around janitor::clean_names() that converts messy column
names into snake_case.
}
\examples{
x <- data.frame("First Name" = 1:3, "AGE " = 4:6, check.names = FALSE)
y <- clean_names(x)
names(y)
}
